refactor: mv everything out of /app that's not a next.js webpage
