+++ frontend/src/app/addContact/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { FormEvent, useState } from "react";
-import ProfileIcon from "../components/profileIcon";
+import ProfileIcon from "@/components/profileIcon";
 import { useRouter } from "next/navigation";
 
 const AddContact = () => {
