+++ frontend/src/utils/arrayUtils.ts
